requirements.txt
llama-index-llms-ollama==0.4.1
pinecone-client==5.0.1
arxiv==2.1.3
llama-index==0.12.2
llama-index-vector-stores-pinecone==0.4.1
llama-index-embeddings-huggingface==0.4.0
python-dotenv=1.0.1